server/auth-service/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-dotenv
motor
